Delete references to messages first
Since we're not able to see who deleted a message, this is a patch to prevent self deletion; as `on_raw_message_delete` checks the database to see if a deleted message in in the database, and the functions that delete messages delete the reference to the message after the message, it's a race condition between on_raw_message_delete firing and the reference being deleted.

To help fix this bug, I moved the reference deletion to before the message is deleted, as `embeds` would be a copy of the list. 

If no messages delete after this, we may need to wrap the embeds query in the list() function first.